+++ pyccel/ast/class_defs.py
@@ -16,7 +16,7 @@ from .builtins   import PythonImag, PythonReal, PythonConjugate
 from .core       import ClassDef, PyccelFunctionDef
 from .datatypes  import (PythonNativeBool, PythonNativeInt, PythonNativeFloat,
                          PythonNativeComplex, StringType, TupleType, CustomDataType,
-                         HomogeneousListType, HomogeneousSetType, DictType)
+                         HomogeneousListType, HomogeneousSetType, DictType, SymbolicType)
 from .numpyext   import (NumpyShape, NumpySum, NumpyAmin, NumpyAmax,
                          NumpyImag, NumpyReal, NumpyTranspose,
                          NumpyConjugate, NumpySize, NumpyResultType, NumpyArray)
@@ -251,7 +251,7 @@ def get_cls_base(class_type):
     NotImplementedError
         Raised if the base class cannot be found.
     """
-    if isinstance(class_type, CustomDataType):
+    if isinstance(class_type, (CustomDataType, SymbolicType)):
         return None
     elif class_type in literal_classes:
         return literal_classes[class_type]